+++ populationsim/balancer.py
@@ -3,7 +3,6 @@
 # See full license in LICENSE.txt.
 
 from builtins import range
-from builtins import object
 import logging
 import numpy as np
 
@@ -26,7 +25,7 @@ MIN_CONTROL_VALUE = 0.1
 MAX_INT = (1 << 31)
 
 
-class ListBalancer(object):
+class ListBalancer:
     """
     Single-geography list balancer using Newton-Raphson method with control relaxation.
