+++ pyproject.toml
@@ -24,6 +24,7 @@ dev = [
     "pytest>=8.3.5",
     "pytest-cov>=5.0.0",
     "ruff>=0.9.9",
+    "vulture>=2.14",
 ]
 
 [pytest.ini]
